feat: getProductList 함수 구현
- 공유되는 요소를 common.js 파일로 이동
